chore: dynamically pack the circleci pipeline configuration (#32618)
* chore: dynamically pack the circleci pipeline configuration

The ./scripts/pack-ci.sh is the first workhorse, here. It uses the circleci
cli to pack any subdirectory of `.circleci/src` into a circleci config
using FYAML format. It can pack every subdirectory, a subsection of
subdirs that include changes, and use circleci cli to validate the
packed yml files.

Husky is set up to validate the packed yml file(s) on commit.

The packed yml file(s) should not be committed to source control. They
are build artifacts. They are configured to be ignored by git in
./.circleci/.gitignore.

./.circleci/config.yml is the second major change. This workflow is
modified such that:
- it uses workflow filtering to prevent the workflow from triggering unless at least
one of the following is true:
  - the workflow was triggered via api (the 'trigger workflows' button in circleci)
  - the branch is `develop`
  - the branch is `release/*`
  - the branch has an open pull request, and the pull request is not draft
- checks out the repository from the HEAD ref of the pr (the branch with the changes),
and calculates a checksum of the contents of ./.circleci/src. Packed workflow files
are cached by the checksum of the source files that were used to generate the packed
workflow. If the cache is restored, the packed pipeline.yml file is continued. If there
is no cache for the checksum, it uses ./scripts/pack-ci to pack the source yml files
and triggers a continuation with the resulting packed yml file.

* rm packed yml file

* revert to old draft detection :( other way is only avail on github app

* just fail instead of cancelling on drafts, more reliable even if not correct

* just halt, since draft->ready does not already trigger a ci run

* circleci agent -> circleci-agent

* validate main circleci yml file; correct entrypoint for pack workflows job

* workflows -> pipeline

* Update pack-ci.sh - newline at eof